// dist/es2015/renderer/shader/chunk/ShaderChunk.d.ts
export declare const empty: GLSLChunk;
export declare const NULL: number;
export declare const basic_materialColor_fragment: GLSLChunk;
export declare const end_basic_fragment: GLSLChunk;
export declare const common_define: GLSLChunk;
export declare const common_fragment: GLSLChunk;
export declare const common_function: GLSLChunk;
export declare const common_vertex: GLSLChunk;
export declare const highp_fragment: GLSLChunk;
export declare const lowp_fragment: GLSLChunk;
export declare const mediump_fragment: GLSLChunk;
export declare const noNormalMap_light_fragment: GLSLChunk;
export declare const lightCommon_fragment: GLSLChunk;
export declare const lightCommon_vertex: GLSLChunk;
export declare const lightEnd_fragment: GLSLChunk;
export declare const light_common: GLSLChunk;
export declare const light_fragment: GLSLChunk;
export declare const light_setWorldPosition_vertex: GLSLChunk;
export declare const light_vertex: GLSLChunk;
export declare const map_forBasic_fragment: GLSLChunk;
export declare const map_forBasic_vertex: GLSLChunk;
export declare const modelMatrix_noInstance_vertex: GLSLChunk;
export declare const normalMatrix_noInstance_vertex: GLSLChunk;
export declare const diffuseMap_fragment: GLSLChunk;
export declare const diffuseMap_vertex: GLSLChunk;
export declare const noDiffuseMap_fragment: GLSLChunk;
export declare const noEmissionMap_fragment: GLSLChunk;
export declare const noLightMap_fragment: GLSLChunk;
export declare const noNormalMap_fragment: GLSLChunk;
export declare const noNormalMap_vertex: GLSLChunk;
export declare const noSpecularMap_fragment: GLSLChunk;
export declare const specularMap_fragment: GLSLChunk;
export declare const specularMap_vertex: GLSLChunk;
export declare const noShadowMap_fragment: GLSLChunk;
export declare type GLSLChunk = {
    top: string;
    define: string;
    varDeclare: string;
    funcDeclare: string;
    funcDefine: string;
    body: string;
};